+++ src/wetterdienst/api.py
@@ -30,6 +30,7 @@ class Wetterdienst:
             "mosmix": "wetterdienst.provider.dwd.mosmix.DwdMosmixRequest",
             "dmo": "wetterdienst.provider.dwd.dmo.DwdDmoRequest",
             "road": "wetterdienst.provider.dwd.road.DwdRoadRequest",
+            "swsmos": "wetterdienst.provider.dwd.swsmos.DwdSwsmosRequest",
             "radar": "wetterdienst.provider.dwd.radar.DwdRadarValues",
             "alerts": "wetterdienst.provider.dwd.alerts.DwdWeatherAlertRequest",
             "derived": "wetterdienst.provider.dwd.derived.DwdDerivedRequest",